pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotlightcurve"
version = "0.1.0"
description = "SpotLightCurve: Analyze stellar activity in exoplanet transit light curves"
readme = "README.md"
authors = [{name = "Arif Solmaz"}]
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "arviz",
    "astropy",
    "astroquery",
    "celerite2",
    "eleanor",
    "jinja2",
    "lightkurve",
    "matplotlib",
    "numpy",
    "pandas",
    "pyyaml",
    "scipy",
    "seaborn",
    "typer[all]",
]

[project.optional-dependencies]
pymc = [
    "pymc",
    "exoplanet",
    "starry",
]
numpyro = [
    "jax[cpu]>=0.4.0",
    "numpyro",
    "tinygp",
]

[project.urls]
Homepage = "https://github.com/arifsolmaz/arifsolmaz.github.io"
Documentation = "https://github.com/arifsolmaz/arifsolmaz.github.io#readme"

[project.scripts]
spotlightcurve = "spotlightcurve.cli:main"
